Apply minor fixes

+++ include/dolphin/os/OSAlloc.h
@@ -26,7 +26,7 @@ OSHeapHandle OSCreateHeap(void *, void *);
 void OSDestroyHeap(size_t idx);
 OSHeapHandle OSSetCurrentHeap(OSHeapHandle);
 void *OSAllocFromHeap(OSHeapHandle, u32);
-long OSCheckHeap(OSHeapHandle);
+size_t OSCheckHeap(OSHeapHandle);
 void OSFreeToHeap(OSHeapHandle heap, void *ptr);
 
 #define OSAlloc(size) OSAllocFromHeap(__OSCurrHeap, (size))
